Extract FolioRecord from FolioReader
