+++ spec/integration/compare_sirsi_and_folio_spec.rb
@@ -21,7 +21,7 @@ describe 'comparing against a well-known location full of documents generated by
 
   shared_examples 'records match' do
     let(:sirsi_record) { sirsi_indexer.map_record(marc_record) }
-    let(:folio_record) { folio_indexer.map_record(Traject::FolioReader::FolioRecord.fetch(catkey)) }
+    let(:folio_record) { folio_indexer.map_record(FolioRecord.fetch(catkey)) }
     let(:mapped_fields) do
       path = File.expand_path('../../lib/traject/config/sirsi_config.rb', __dir__)
       File.read(path).scan(/to_field ["']([^"']+)["']/).map(&:first).uniq
